[US301] Done, except weird Edge syntax
Still double running

+++ ESINF/src/main/java/app/domain/model/ClientsProducers.java
@@ -8,11 +8,11 @@ public class ClientsProducers {
 
     private String locId;
 
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
     private String code;
 
-    public ClientsProducers(String locId, double latitude, double longitude, String code) {
+    public ClientsProducers(String locId, float latitude, float longitude, String code) {
         this.locId = locId;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -39,11 +39,11 @@ public class ClientsProducers {
         return locId;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public double getLongitude() {return longitude;}
+    public float getLongitude() {return longitude;}
 
     public String getCode() {
         return code;
